+++ bdc_catalog/models/__init__.py
@@ -10,7 +10,8 @@
 
 from .band import Band, BandSRC
 from .base_sql import db
-from .collection import Collection, CollectionsProviders, CollectionSRC
+from .collection import (Collection, CollectionRole, CollectionsProviders,
+                         CollectionSRC)
 from .composite_function import CompositeFunction
 from .grid_ref_sys import GridRefSys
 from .item import Item, ItemsProcessors, SpatialRefSys
@@ -19,6 +20,7 @@ from .processor import Processor
 from .provider import Provider
 from .quicklook import Quicklook
 from .resolution_unit import ResolutionUnit
+from .role import Role
 from .tile import Tile
 from .timeline import Timeline
 
@@ -27,6 +29,7 @@ __all__ = (
     'Band',
     'BandSRC',
     'Collection',
+    'CollectionRole',
     'CollectionSRC',
     'CollectionsProviders',
     'CompositeFunction',
@@ -38,6 +41,7 @@ __all__ = (
     'Provider',
     'Quicklook',
     'ResolutionUnit',
+    'Role',
     'SpatialRefSys',
     'Tile',
     'Timeline',